fix: return null for errors in get
The redis cache is a convenient performance improvement, but should not inhibit the function of the
underlying memoized function. An error fetching from Redis or parsing the result should return null
rather than throwing. This preserves the underlying functionality. It also allows bad values in the
cache to be overwritten as a null returned from `get` will cause that key to be overwritten.